+++ lib/Model/dayselector.dart
@@ -1,5 +1,17 @@
 import 'package:flutter/material.dart';
 
+import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:dropdown_search/dropdown_search.dart';
+import "./search.dart";
+import '../UI/scheduling_interface.dart';
+import 'package:dropdown_search/dropdown_search.dart';
+import 'package:firebase_database/firebase_database.dart';
+import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
+import 'package:google_fonts/google_fonts.dart';
+
 class DaySelect extends StatefulWidget {
   DaySelect({Key key}) : super(key: key);
 
@@ -7,73 +19,164 @@ class DaySelect extends StatefulWidget {
   _DaySelectState createState() => _DaySelectState();
 }
 
-class Day {
-  int id;
-  String name;
-  Day(this.id, this.name);
-
-  static List<Day> getDays() {
-    return <Day>[
-      Day(1, 'Sunday'),
-      Day(2, 'Monday'),
-      Day(3, 'Tuesday'),
-      Day(4, 'Wednesday'),
-      Day(5, 'Thursday'),
-      Day(6, 'Friday'),
-      Day(7, 'Saturday'),
-    ];
-  }
-}
-
 class _DaySelectState extends State<DaySelect> {
-  List<Day> _days = Day.getDays();
-  List<DropdownMenuItem<Day>> _dropdownMenuItems;
-  Day _selectedDay;
-
-  @override
-  void initState() {
-    _dropdownMenuItems = buildDropDownMenuItems(_days);
-    _selectedDay = _dropdownMenuItems[0].value;
-    super.initState();
-  }
-
-  List<DropdownMenuItem<Day>> buildDropDownMenuItems(List days) {
-    List<DropdownMenuItem<Day>> items = List();
-    for (Day day in days) {
-      items.add(DropdownMenuItem(
-        value: day,
-        child: Text(day.name),
-      ));
-    }
-    return items;
-  }
-
-  onChangedDropdownItem(Day selectedDay) {
-    setState(() {
-      _selectedDay = selectedDay;
-    });
-  }
-
   @override
   Widget build(BuildContext context) {
+    String daySelected = "";
     return Container(
-      margin: EdgeInsets.only(right: 30),
-      decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(30),
-          border: Border.all(),
-          color: Color(0xFF398AE5)),
-      child: Padding(
-        padding: const EdgeInsets.only(right: 10, left: 10),
-        child: DropdownButtonHideUnderline(
-          child: DropdownButton(
-            value: _selectedDay,
-            items: _dropdownMenuItems,
-            onChanged: onChangedDropdownItem,
-            icon: Icon(Icons.keyboard_arrow_down),
-            iconSize: 25,
+      child: Column(children: [
+        Container(
+          decoration: BoxDecoration(
+              color: Colors.white, borderRadius: BorderRadius.circular(20)),
+          child: DropdownSearch<String>(
+            mode: Mode.BOTTOM_SHEET,
+            maxHeight: 300,
+            items: [
+              "Sunday",
+              "Monday",
+              "Tuesday",
+              "Wednesday",
+              "Thursday",
+              "Friday",
+              "Saturday"
+            ],
+            dropDownButton: DropdownButton(
+              icon: Icon(Icons.arrow_circle_down),
+              iconDisabledColor: Color(0xFF398AE5),
+              iconSize: 30,
+            ),
+            dropdownSearchDecoration: InputDecoration(
+              hintText: "Select a Day!",
+              hintStyle: TextStyle(
+                  color: Color(0xFF398AE5),
+                  letterSpacing: 1,
+                  fontSize: 15,
+                  fontWeight: FontWeight.bold),
+              enabledBorder: OutlineInputBorder(
+                  borderSide: BorderSide(color: Color(0xFF398AE5)),
+                  borderRadius: BorderRadius.circular(20)),
+              contentPadding: EdgeInsets.fromLTRB(12, 12, 8, 0),
+            ),
+            popupBarrierColor: Colors.white,
+            onChanged: (day) {
+              daySelected = day;
+              //Method to fetch common empty slots for the day as soon as the user selects a day
+              fetchEmptySlots(daySelected);
+            },
+            popupBackgroundColor: Colors.white,
+            showClearButton: true,
+            showSearchBox: true,
+            searchBoxDecoration: InputDecoration(
+              border:
+                  OutlineInputBorder(borderRadius: BorderRadius.circular(10)),
+              fillColor: Colors.white,
+              contentPadding: EdgeInsets.fromLTRB(12, 12, 8, 0),
+              labelText: "Search the day!",
+            ),
+            popupTitle: Container(
+              height: 50,
+              decoration: BoxDecoration(
+                color: Color(0xFF398AE5),
+                borderRadius: BorderRadius.only(
+                  topLeft: Radius.circular(20),
+                  topRight: Radius.circular(20),
+                ),
+              ),
+              child: Center(
+                child: Text(
+                  'Conduct a meet on?',
+                  style: TextStyle(
+                    fontSize: 15,
+                    fontWeight: FontWeight.bold,
+                    color: Colors.white,
+                  ),
+                ),
+              ),
+            ),
+            popupShape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.only(
+                topLeft: Radius.circular(24),
+                topRight: Radius.circular(24),
+              ),
+            ),
           ),
         ),
-      ),
+        SizedBox(
+          height: 40,
+        ),
+        Container(
+            decoration: BoxDecoration(
+                color: Colors.white, borderRadius: BorderRadius.circular(20)),
+            child: ExpansionTile(
+              maintainState: true,
+              title: Text(
+                "Choose Slots",
+                style: GoogleFonts.aBeeZee(
+                    textStyle:
+                        TextStyle(color: Color(0xFF398AE5), fontSize: 15),
+                    letterSpacing: 1,
+                    fontWeight: FontWeight.bold),
+              ),
+            ))
+      ]),
     );
   }
+
+  void fetchEmptySlots(String daySelected) async {
+    print("selected Uid: $selectedNames");
+    print("Day selected: $daySelected");
+    int test = selectedNames.elementAt(0).trim().length;
+    print("test: $test");
+    //Declaring the firebase instances and references
+    FirebaseDatabase databaseSlots = new FirebaseDatabase();
+    DatabaseReference referenceSlots =
+        databaseSlots.reference().child("schedule");
+    List<dynamic> referenceSchedule = [];
+    var recurringSet = <int>{};
+    var tempSet = <int>{};
+    if (totalSelected != 0) {
+      for (int i = 0; i < selectedNames.length; i++) {
+        if (i == 0) {
+          await referenceSlots
+              .child(selectedNames.elementAt(i).trim())
+              .child(daySelected)
+              .once()
+              .then((DataSnapshot dataSnapshot1) {
+            referenceSchedule = dataSnapshot1.value;
+            print("Schedule: $referenceSchedule");
+            for (int j = 0; j < referenceSchedule.length; j++) {
+              if (referenceSchedule.elementAt(j) == "true") {
+                recurringSet.add(j);
+              }
+            }
+            print("individual Set: $recurringSet");
+          });
+        } else {
+          await referenceSlots
+              .child(selectedNames.elementAt(i).trim())
+              .child(daySelected)
+              .once()
+              .then((DataSnapshot dataSnapshot1) {
+            referenceSchedule = dataSnapshot1.value;
+            print("Schedule: $referenceSchedule");
+          });
+          for (int j = 0; j < referenceSchedule.length; j++) {
+            if (referenceSchedule.elementAt(j) == "true") {
+              tempSet.add(j);
+            }
+          }
+          recurringSet = tempSet.intersection(recurringSet);
+          tempSet.clear();
+          print("individual Set: $recurringSet");
+        }
+      }
+      print("Commom FreeSlots: $recurringSet");
+    } else {
+      Fluttertoast.showToast(
+          msg: "Please select atleast one member for the meeting");
+    }
+    displaySlots(recurringSet);
+  }
+
+  Widget displaySlots(Set recurringSet) {}
 }
